+++ tests/test_utils/test_validators.py
@@ -8,7 +8,7 @@ from django.core.exceptions import ValidationError
 from graphene_django_extensions.testing.utils import parametrize_helper
 
 from tilavarauspalvelu.typing import TimeSlot
-from tilavarauspalvelu.utils.validators import validate_reservable_times_begin_end, validate_reservable_times_overlap
+from tilavarauspalvelu.validators import validate_reservable_times_begin_end, validate_reservable_times_overlap
 
 if TYPE_CHECKING:
     from collections.abc import Collection
